requirements.txt
# Binance API library for interacting with the Binance platform
python-binance==1.0.16

# Library for writing and running automated tests
pytest==8.3.5

# Mocking library for unit tests
pytest-mock==3.10.0

# Library for reading and working with configuration files
configparser==5.3.0

# Library for data manipulation and analysis
pandas==2.0.1

# Library for numerical computations
numpy==1.24.3

# Library for creating visualizations and plots
matplotlib==3.7.1

# Requests library for HTTP requests
requests==2.31.0

# Logging library for structured logging
loguru==0.7.0
